move test and lib files
