+++ test/near_earth_objects_test.rb
@@ -1,7 +1,7 @@
 require 'minitest/autorun'
 require 'minitest/pride'
 require 'pry'
-require_relative 'near_earth_objects'
+require_relative '../lib/near_earth_objects'
 
 class NearEarthObjectsTest < Minitest::Test
   def test_a_date_returns_a_list_of_neos
